+++ discordrpc/__init__.py
@@ -5,6 +5,7 @@ from .exceptions import (
     RPCException, Error, DiscordNotOpened, ActivityError,
     InvalidURL, InvalidID, ButtonError, ProgressbarError,
     InvalidActivityType, ActivityTypeDisabled,
+    InvalidEvent, InvalidEventType,
 )
 from .types import Activity, StatusDisplay, User, Application, Event
 from .utils import remove_none, timestamp, date_to_timestamp, use_local_time, progress_bar, get_app_info
@@ -13,7 +14,7 @@ __title__ = "Discord RPC"
 try:
     __version__ = _pkg_ver('discord-rpc')
 except PackageNotFoundError:
-    __version__ = "6.5b1"
+    __version__ = "6.5b2"
 __authors__ = "Senophyx"
 __license__ = "MIT License"
 __copyright__ = "Copyright 2021-2025 Senophyx"